+++ bilibili_api/channel.py
@@ -3,16 +3,13 @@ bilibili_api.channel
 
 频道相关，与视频分区不互通。
 """
-import re
-import json
 from enum import Enum
 from typing import List, Union, Optional
 
 from bilibili_api.utils.utils import get_api
 from bilibili_api.utils.credential import Credential
 from bilibili_api.utils.initial_state import get_initial_state
-from bilibili_api.errors import ApiException, ResponseException
-from bilibili_api.utils.network_httpx import request, Api
+from bilibili_api.utils.network_httpx import Api
 
 API = get_api("channel")
