+++ bilibili_api/channel_series.py
@@ -11,7 +11,7 @@ import httpx
 
 from .utils.utils import get_api
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API_USER = get_api("user")
 API = get_api("channel-series")
