+++ bilibili_api/comment.py
@@ -17,7 +17,7 @@ from typing import Union
 
 from .utils.utils import get_api
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 from .exceptions.ArgsException import ArgsException
 
 API = get_api("common")
