+++ bilibili_api/creative_center.py
@@ -11,7 +11,7 @@ from typing import List, Union
 
 from .utils.utils import get_api
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API = get_api("creative_center")
