+++ bilibili_api/favorite_list.py
@@ -11,7 +11,7 @@ from . import user
 from .video import Video
 from .utils.utils import join, get_api
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 from .exceptions.ArgsException import ArgsException
 
 API = get_api("favorite-list")
