+++ bilibili_api/game.py
@@ -8,7 +8,7 @@ from typing import Union
 
 from .utils.utils import get_api
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API = get_api("game")
