+++ bilibili_api/interactive_video.py
@@ -25,7 +25,7 @@ from .video import Video
 from .utils.utils import get_api
 from .utils.AsyncEvent import AsyncEvent
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API = get_api("interactive_video")
