+++ bilibili_api/live_area.py
@@ -11,7 +11,7 @@ from enum import Enum
 from typing import Dict, List, Tuple, Union
 
 from .utils.utils import get_api
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API = get_api("live-area")
