+++ bilibili_api/login.py
@@ -29,7 +29,7 @@ from .utils.utils import get_api
 from .utils.credential import Credential
 from .exceptions.LoginError import LoginError
 from .utils.network import to_form_urlencoded
-from .utils.network_httpx import HEADERS, get_session, get_spi_buvid_sync
+from .utils.network import HEADERS, get_session, get_spi_buvid_sync
 from .utils.captcha import get_result, close_server, start_server
 from .utils.safecenter_captcha import get_result as safecenter_get_result
 from .utils.safecenter_captcha import close_server as safecenter_close_server
