+++ bilibili_api/session.py
@@ -21,7 +21,7 @@ from .utils.utils import get_api
 from .utils.picture import Picture
 from .utils.AsyncEvent import AsyncEvent
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API = get_api("session")
