+++ bilibili_api/topic.py
@@ -11,7 +11,7 @@ from . import dynamic
 from .user import get_self_info
 from .utils.utils import get_api
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API = get_api("topic")
