+++ bilibili_api/video_tag.py
@@ -11,7 +11,7 @@ import httpx
 from .errors import *
 from .utils.utils import get_api
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API = get_api("video_tag")
 API_video = get_api("video")
