+++ bilibili_api/vote.py
@@ -11,7 +11,7 @@ from typing import Union, Optional
 from .utils.utils import get_api
 from .utils.picture import Picture
 from .utils.credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 
 API = get_api("vote")
