+++ bilibili_api/watchroom.py
@@ -10,7 +10,7 @@ from enum import Enum
 from typing import Dict, List, Union
 
 from .credential import Credential
-from .utils.network_httpx import Api
+from .utils.network import Api
 from .utils.utils import get_api
 
 API = get_api("watchroom")
